fix(なし): 🐞 DB側で小数点以下切り捨てられる問題

+++ supabase_schema.sql
@@ -35,7 +35,7 @@ CREATE TABLE saved_tags (
 );
 
 ALTER TABLE todos ADD COLUMN order_index INTEGER DEFAULT 0;
-ALTER TABLE todos ADD COLUMN estimated_hours INTEGER;
+ALTER TABLE todos ADD COLUMN estimated_hours NUMERIC(5,2);
 ALTER TABLE todos ADD COLUMN recurrence_rule TEXT;
 
 -- 4. RLS (Row Level Security) の有効化
